Fixed some warnings in SamplePosition

+++ src/rec/base/SamplePosition.h
@@ -9,8 +9,9 @@ struct SamplePosition {
   SamplePosition() : position_(0) {}
   SamplePosition(int64 p) : position_(p) {}
   SamplePosition(int p) : position_(p) {}
-  SamplePosition(RealTime time) : position_(44100 * time) {}
-  SamplePosition(float time) : position_(44100 * time) {}
+  SamplePosition(RealTime time) : position_(static_cast<int64>(44100 * time)) {}
+  SamplePosition(float time) : position_(static_cast<int64>(44100 * time)) {}
+
   SamplePosition& operator++() { ++position_; return *this; }
   SamplePosition& operator--() { --position_; return *this; }
